Libvirt: fix libvirt_use_sudo option
When parsing the config file, if libvirt_use_sudo is present and set to
a string that represents a boolean, then the yaml parser will not give
us a String object, but something from TrueClass or FalseClass.

Since we need a string later on (otherwise, the tests on the variable
always fails, even if libvirt_use_sudo is true), convert this to a
string.