Fix: Configure host network before configuring nodes
If the node network is not connected to any physical networks, its
bridge might not exist if no vifs are attached to it. In these
situations, the gateway's IP config need to be performed after the
bridge exists.

This fix makes sure that the gateway's IP configuration takes places
before each node's network configuration.